Working on the repository layer
I really kind of want to use prepared statements but my first foray into them with select statements is not good.  I haven't done this low of programming in A LONG TIME and I've been drinking a bit so the unused/commented out code could indeed be insane.  Maybe that's why it's not working?  I don't know... doesn't seem right though.

Commiting this stuff just to have a record of it and the next commit is cleaning up the unused/definitely won't be used stuff up.  Already I'm trying to make the amount of boilerplate I write non-existent... kind of tough when you don't even fully know how to speak the language ;).